new tests in misc/arrays.py
